+++ exchange/src/main/java/com/test/exchange/exchange/service/ExchangeService.java
@@ -6,19 +6,17 @@ import com.test.exchange.exchange.domain.Currency;
 import com.test.exchange.exchange.domain.Exchange;
 import com.test.exchange.exchange.domain.Remittance;
 import com.test.exchange.exchange.dto.ExchangeMoneyResponse;
-import com.test.exchange.exchange.repository.ExchangeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class ExchangeService {
-    private final ExchangeRepository exchangeRepository;
     private final ExchangeRateClient exchangeRateClient;
+    private final ExchangeRepoClient exchangeRepoClient;
 
     @Value("${api-layer.base-url}")
     private String baseUrl;
@@ -45,16 +43,9 @@ public class ExchangeService {
     public ExchangeMoneyResponse exchangeMoney(Currency source, Currency target, double givenRemittance){
         Remittance remittance = Remittance.from(givenRemittance);
         double exchangeRate = getExchangeRate(source,target);
-
-        Exchange exchange = saveExchange(source,target,remittance,exchangeRate);
+        Exchange exchange = exchangeRepoClient.saveExchange(source,target,remittance,exchangeRate);
 
         return new ExchangeMoneyResponse(exchange.calculateRemittance(),target);
-    }
-
-    @Transactional
-    public Exchange saveExchange(Currency source, Currency target, Remittance remittance, double exchangeRate){
-        Exchange exchange = Exchange.of(source,target,remittance,exchangeRate);
-        return exchangeRepository.save(exchange);
     }
 
 }
